pyproject.toml
[project]
name = "habits-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "volko",email = "Toxa.Vol@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=2.2,<6.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "celery (>=5.5.1,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "eventlet (>=0.39.1,<0.40.0)",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
flake8 = "^7.2.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.0"
pillow = "^11.2.1"
coverage = "^7.8.0"

